--- Assets/Scripts/SS3D/Systems/Inventory/Filters.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a1f075f2d7b66c449865d14a496f5d7b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
