+++ Assets/Scripts/SS3D/Systems/Storage/Containers/Editor/ContainerDescriptorEditor.cs
@@ -3,6 +3,7 @@ using UnityEditor;
 using UnityEngine;
 using SS3D.Systems.Storage.Containers;
 using FishNet.Object;
+using SS3D.Systems;
 
 
 /// <summary>
