+++ Assets/Scripts/SS3D/Systems/Traits.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c5cc1e30487a45044a89ebee90a6bb36
+guid: 3868d24c63cd3f64d899de4ecd02c0ed
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
