+++ Assets/Scripts/SS3D/Systems/Traits/Filter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a30489427894cdc4cbbbda4e09cb68e8
+guid: 0adc192349dc57848b5394a2c6707c23
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
