+++ Assets/Scripts/SS3D/Systems/Traits/FilterEnums.cs
