+++ Assets/Scripts/SS3D/Systems/Traits/FilterEnums.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 192f2453d05363c4bb6101f58fb7c844
+guid: 2bf3383363bd491418f5b43a0c27c92a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
