+++ Assets/Scripts/SS3D/Systems/Traits/Trait.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2338fa1b64ca78e4680e5d088e153d4a
+guid: b120a07bf55051644adda51d564ace97
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
